Don't require signin for market page
