+++ package.json
@@ -14,37 +14,37 @@
     "check:all": "npm run lint && npm run check:format && npm run check:types"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.12",
+    "@headlessui/react": "^1.7.13",
     "@heroicons/react": "^2.0.16",
     "@next-auth/prisma-adapter": "1.0.5",
     "@prisma/client": "^4.11.0",
-    "@tanstack/react-query": "^4.24.10",
+    "@tanstack/react-query": "^4.26.0",
     "@tremor/react": "^1.8.1",
-    "@trpc/client": "^10.13.2",
-    "@trpc/next": "^10.13.2",
-    "@trpc/react-query": "^10.13.2",
-    "@trpc/server": "^10.13.2",
+    "@trpc/client": "^10.14.0",
+    "@trpc/next": "^10.14.0",
+    "@trpc/react-query": "^10.14.0",
+    "@trpc/server": "^10.14.0",
     "axios": "^1.3.4",
     "clsx": "^1.2.1",
-    "next": "13.2.1",
-    "next-auth": "4.19.2",
+    "next": "13.2.3",
+    "next-auth": "4.20.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "superjson": "1.12.2",
-    "zod": "^3.20.6"
+    "zod": "^3.21.2"
   },
   "devDependencies": {
-    "@types/node": "^18.14.2",
+    "@types/node": "^18.14.6",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.54.0",
     "@typescript-eslint/parser": "^5.54.0",
     "autoprefixer": "^10.4.13",
     "eslint": "^8.35.0",
-    "eslint-config-next": "^13.2.1",
+    "eslint-config-next": "^13.2.3",
     "postcss": "^8.4.21",
     "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.3",
+    "prettier-plugin-tailwindcss": "^0.2.4",
     "prisma": "^4.11.0",
     "tailwindcss": "^3.2.7",
     "typescript": "^4.9.5"
