+++ src/components/Layout/links.ts
@@ -1,6 +1,6 @@
 export const navLinks = [
-  { name: "Dashboard", href: "/" },
+  { name: "Market", href: "/" },
+  { name: "Dashboard", href: "/dashboard" },
   { name: "Transactions", href: "/transactions" },
-  { name: "Market", href: "/market" },
   { name: "Trade", href: "/trade" },
 ];
